[chore]: performed squash on prev 44 commits.
This squash was performed to clean up cluttered commits and commit messages.
All the open issues and PR's pre-squash, are fixed in this commit.
As a result to squash, previous releases from github are deleted.

This commit contains the code for v2.3.0
The readme section provides all functionalities, this extension provides.
Finally, all issues will be tracked from the github issue webpage.

-kalpaj agrawalla